refactor: share the workspace-stamping audit recorder across MCP tool modules
